put unit tests in a new file where they belong
